FGD/requirements.txt
discord.py
selenium
python-dotenv
pillow
